src/tools/imageClassification/requirements.txt
torch
torchvision
scikit-learn
pandas
tensorflow
matplotlib
redis
python-dotenv
numpy
redisearch
tqdm